+++ docs/release-notes.md
@@ -15,6 +15,7 @@ pyhub.mcptools update
 
 ## v0.9.3
 
++ macOS에 피봇 테이블 생성 지원 추가
 + 간략하게 `pyhub.mcptools` 외에 `mcptools` 동일한 명령 추가
 
 ## v0.9.2